+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Fixes
 
-- Start SDK on the main thread (#3291)
+- Always start SDK on the main thread (#3291)
 - App hang with race condition for tick counter (#3290)
 - Remove "duplicate library" warning (#3312)
 - Remove unnecessary build settings (#3325)